# prism/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API gateway with service routing, audit logging, API key management, and service discovery.",
        "title": "Prism API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/apikey": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated list of API keys filtered by tenant or user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikey"
                ],
                "summary": "List API keys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ApiKeyResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new API key for a tenant or user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikey"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "API key creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ApiKeyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ApiKeyCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/apikey/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validate an API key and return its details if valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikey"
                ],
                "summary": "Validate API key",
                "parameters": [
                    {
                        "description": "API key validation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ApiKeyValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/apikey/{apikeyID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve an API key by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikey"
                ],
                "summary": "Get API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key ID",
                        "name": "apikeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke an existing API key by ID",
                "tags": [
                    "apikey"
                ],
                "summary": "Revoke API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key ID",
                        "name": "apikeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
        "/apikey/{apikeyID}/metadata": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the metadata of an existing API key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikey"
                ],
                "summary": "Update API key metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key ID",
                        "name": "apikeyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Metadata update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateApiKeyMetadataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid JSON request"
                },
                "slug": {
                    "type": "string",
                    "example": "invalid-json"
                }
            }
        },
        "ApiKeyCreateRequest": {
            "type": "object",
            "properties": {
                "expires_in_days": {
                    "type": "integer",
                    "example": 30
                },
                "name": {
                    "type": "string",
                    "example": "my-api-key"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read:users",
                        "write:users"
                    ]
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "ApiKeyCreateResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "key_prefix": {
                    "type": "string",
                    "example": "ak_abcdefgh"
                },
                "name": {
                    "type": "string",
                    "example": "my-api-key"
                },
                "raw_api_key": {
                    "type": "string",
                    "example": "ak_abcdefghijklmnopqrstuvwxyz123456"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read:users",
                        "write:users"
                    ]
                }
            }
        },
        "ApiKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_ip_address": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "key_prefix": {
                    "type": "string",
                    "example": "ak_abcdefgh"
                },
                "last_used_at": {
                    "type": "string"
                },
                "last_used_ip_address": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "name": {
                    "type": "string",
                    "example": "my-api-key"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read:users",
                        "write:users"
                    ]
                },
                "tenant_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                }
            }
        },
        "ApiKeyValidateRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "ak_abc123..."
                }
            }
        },
...
        "UpdateApiKeyMetadataRequest": {
            "type": "object",
            "properties": {
                "expires_in_days": {
                    "type": "integer",
                    "example": 90
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "updated-key-name"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}